Fix news item title layout issue
We need to consider the title button's title label height in order to
set the height of the button correctly. Right now we're just forcing a
`-layoutIfNeeded` in order to ascertain the height of the label, and
then tweaking the button height constraint with the calculated value.

Not sure if there's a cleaner way of doing this, but it'll do for now!

[Finishes #106247256]