Tweak zip code validation implementation
Having spoken with design, their preferred behavior is that rather than
showing an alert for an invalid zip code, we disable the "Search" button
instead.

In addition to this behavior change, I've converted the specs to Quick
for consistency and also introduced a ZipCodeValidator protocol so that
we can have a purely fake validator under test. We also now inject the
validator into the events controller for ease of testing.