Improve action alert rendering
The only issue that we have in production content right now is that
Tweets have ~20px of additional bottom padding the first time they are
rendered. Subsequent views (i.e., swipe away and back) self-correct.

This isn't an issue yet, but if we need to add buttons below the
UIWebView, then we'll need to fix this for sure.